// code/src/interfaces/eventOnSica3NuevoIngreso.ts
// Generated by https://quicktype.io

export interface InterfaceNuevoIngresoHospitalario {
    id_paciente:     number;
    tipo:            string;
    dx_ingreso:      string;
    via_ingreso:     number;
    codigo_atencion: string[];
    id_especialidad: number;
    id_area:         number;
    id_medico:       number;
    id_cama:         number;
    tipo_ingreso:    string;
    clasificacion: string;
    accion:        string;
}