// code/src/interfaces/eventOnZumbido.ts
// Generated by https://quicktype.io

export interface InterfaceZumbido {
    data_emisor: DataEmisor;
    id_receptor: number;
}

export interface DataEmisor {
    id_usuario:      number;
    nombre_completo: string;
    sexo:            string;
    id_empleado:     number;
    matricula:       string;
    puesto:          string;
    permisos:        Permisos;
    foto_portada:    null;
    foto_perfil:     null;
}

export interface Permisos {
    roles:          string;
    lista_permisos: null[];
}